refactor(llm): Extract shared utilities and remove langchain4j integration
- Extract common parsing and reference handling functions to shared-utils.ts
- Remove langchain4j-integration module (moved to separate repository)
- Remove ModelCapabilityService and related tests
- Remove demo.sh script
- Update OpenAPI spec and related files
