+++ app/llm-artifact-types-src/src/PromptTemplateArtifactType.ts
@@ -11,12 +11,15 @@ import type {
     ReferenceFinderRequest,
     ReferenceFinderResponse
 } from '@apicurio/artifact-type-builtins';
-import { parse as parseYaml, stringify as stringifyYaml } from 'yaml';
-
-type TypedContent = {
-    contentType: string;
-    content: string;
-}
+import { stringify as stringifyYaml } from 'yaml';
+import {
+    TypedContent,
+    parseContent,
+    sortObjectKeys,
+    findRefs,
+    resolveRefs,
+    rewriteRefsWithUrls
+} from './shared-utils';
 
 interface VariableSchema {
     type?: string;
@@ -39,29 +42,6 @@ interface PromptTemplate {
     variables?: Record<string, VariableSchema>;
     outputSchema?: Record<string, unknown>;
     metadata?: Record<string, unknown>;
-}
-
-/**
- * Parse content as JSON or YAML based on content type
- */
-function parseContent(content: string, contentType: string): PromptTemplate | null {
-    try {
-        if (contentType === 'application/json') {
-            return JSON.parse(content);
-        } else if (contentType === 'application/x-yaml' ||
-                   contentType === 'application/yaml' ||
-                   contentType === 'text/x-prompt-template') {
-            return parseYaml(content) as PromptTemplate;
-        }
-        // Try JSON first, then YAML
-        try {
-            return JSON.parse(content);
-        } catch {
-            return parseYaml(content) as PromptTemplate;
-        }
-    } catch {
-        return null;
-    }
 }
 
 /**
@@ -94,7 +74,7 @@ export function acceptsContent(request: ContentAccepterRequest): boolean {
         return false;
     }
 
-    const parsed = parseContent(content, contentType);
+    const parsed = parseContent<PromptTemplate>(content, contentType);
     if (!parsed) {
         return false;
     }
@@ -133,7 +113,7 @@ export function validate(request: ContentValidatorRequest): ContentValidatorResp
     }
 
     // Parse content
-    const parsed = parseContent(content, contentType);
+    const parsed = parseContent<PromptTemplate>(content, contentType);
     if (!parsed) {
         violations.push({
             description: 'Invalid content format. Unable to parse as JSON or YAML.',
@@ -250,8 +230,8 @@ export function testCompatibility(request: CompatibilityCheckerRequest): Compati
     const incompatibleDifferences: Array<{ description: string; context: string }> = [];
 
     try {
-        const proposed = parseContent(proposedArtifact.content, proposedArtifact.contentType);
-        const existing = parseContent(existingArtifacts[0].content, existingArtifacts[0].contentType);
+        const proposed = parseContent<PromptTemplate>(proposedArtifact.content, proposedArtifact.contentType);
+        const existing = parseContent<PromptTemplate>(existingArtifacts[0].content, existingArtifacts[0].contentType);
 
         if (!proposed || !existing) {
             return {
@@ -356,24 +336,6 @@ export function testCompatibility(request: CompatibilityCheckerRequest): Compati
 }
 
 /**
- * Sort object keys recursively for canonicalization
- */
-function sortObjectKeys(obj: unknown): unknown {
-    if (obj === null || typeof obj !== 'object') {
-        return obj;
-    }
-    if (Array.isArray(obj)) {
-        return obj.map(sortObjectKeys);
-    }
-    const sorted: Record<string, unknown> = {};
-    const keys = Object.keys(obj as Record<string, unknown>).sort();
-    for (const key of keys) {
-        sorted[key] = sortObjectKeys((obj as Record<string, unknown>)[key]);
-    }
-    return sorted;
-}
-
-/**
  * Canonicalize PROMPT_TEMPLATE content by normalizing YAML structure
  */
 export function canonicalize(request: ContentCanonicalizerRequest): ContentCanonicalizerResponse {
@@ -381,7 +343,7 @@ export function canonicalize(request: ContentCanonicalizerRequest): ContentCanon
     const contentType: string = request.content.contentType;
 
     try {
-        const parsed = parseContent(content, contentType);
+        const parsed = parseContent<PromptTemplate>(content, contentType);
         if (!parsed) {
             return {
                 typedContent: {
@@ -418,32 +380,6 @@ export function canonicalize(request: ContentCanonicalizerRequest): ContentCanon
 }
 
 /**
- * Find all $ref references in variable schemas
- */
-function findRefs(obj: unknown, refs: string[]): void {
-    if (obj === null || typeof obj !== 'object') {
-        return;
-    }
-    if (Array.isArray(obj)) {
-        for (const item of obj) {
-            findRefs(item, refs);
-        }
-        return;
-    }
-    const record = obj as Record<string, unknown>;
-    if (typeof record.$ref === 'string') {
-        const ref = record.$ref;
-        // Only add external references (not internal #/...)
-        if (!ref.startsWith('#') && !refs.includes(ref)) {
-            refs.push(ref);
-        }
-    }
-    for (const key of Object.keys(record)) {
-        findRefs(record[key], refs);
-    }
-}
-
-/**
  * Find external references in the PROMPT_TEMPLATE content
  */
 export function findExternalReferences(request: ReferenceFinderRequest): ReferenceFinderResponse {
@@ -453,7 +389,7 @@ export function findExternalReferences(request: ReferenceFinderRequest): Referen
     const references: string[] = [];
 
     try {
-        const parsed = parseContent(content, contentType);
+        const parsed = parseContent<PromptTemplate>(content, contentType);
         if (parsed) {
             // Look for $ref in variables and outputSchema
             if (parsed.variables) {
@@ -468,38 +404,6 @@ export function findExternalReferences(request: ReferenceFinderRequest): Referen
     }
 
     return { externalReferences: references };
-}
-
-/**
- * Resolve $ref references in an object recursively
- */
-function resolveRefs(obj: unknown, indexedRefs: Record<string, TypedContent>): unknown {
-    if (obj === null || typeof obj !== 'object') {
-        return obj;
-    }
-    if (Array.isArray(obj)) {
-        return obj.map(item => resolveRefs(item, indexedRefs));
-    }
-    const record = obj as Record<string, unknown>;
-    if (typeof record.$ref === 'string') {
-        const ref = record.$ref;
-        // Resolve external references
-        if (!ref.startsWith('#') && indexedRefs[ref]) {
-            try {
-                const resolved = parseContent(indexedRefs[ref].content, indexedRefs[ref].contentType);
-                if (resolved) {
-                    return resolved;
-                }
-            } catch {
-                // Keep original ref if parse fails
-            }
-        }
-    }
-    const result: Record<string, unknown> = {};
-    for (const key of Object.keys(record)) {
-        result[key] = resolveRefs(record[key], indexedRefs);
-    }
-    return result;
 }
 
 /**
@@ -518,7 +422,7 @@ export function dereference(request: ContentDereferencerRequest): ContentDerefer
     });
 
     try {
-        const parsed = parseContent(content, contentType);
+        const parsed = parseContent<PromptTemplate>(content, contentType);
         if (!parsed) {
             return {
                 typedContent: {
@@ -528,7 +432,7 @@ export function dereference(request: ContentDereferencerRequest): ContentDerefer
             };
         }
 
-        const dereferenced = resolveRefs(parsed, indexedResolvedRefs);
+        const dereferenced = resolveRefs(parsed, indexedResolvedRefs, parseContent<PromptTemplate>);
         const dereferencedContent = stringifyYaml(dereferenced, {
             lineWidth: 0,
             defaultKeyType: 'PLAIN',
@@ -552,33 +456,6 @@ export function dereference(request: ContentDereferencerRequest): ContentDerefer
 }
 
 /**
- * Rewrite $ref values to use resolved URLs
- */
-function rewriteRefsWithUrls(obj: unknown, indexedUrls: Record<string, string>): unknown {
-    if (obj === null || typeof obj !== 'object') {
-        return obj;
-    }
-    if (Array.isArray(obj)) {
-        return obj.map(item => rewriteRefsWithUrls(item, indexedUrls));
-    }
-    const record = obj as Record<string, unknown>;
-    const result: Record<string, unknown> = {};
-    for (const key of Object.keys(record)) {
-        if (key === '$ref' && typeof record[key] === 'string') {
-            const ref = record[key] as string;
-            if (!ref.startsWith('#') && indexedUrls[ref]) {
-                result[key] = indexedUrls[ref];
-            } else {
-                result[key] = ref;
-            }
-        } else {
-            result[key] = rewriteRefsWithUrls(record[key], indexedUrls);
-        }
-    }
-    return result;
-}
-
-/**
  * Rewrite references to use URLs instead of file paths
  */
 export function rewriteReferences(request: ContentDereferencerRequest): ContentDereferencerResponse {
@@ -591,7 +468,7 @@ export function rewriteReferences(request: ContentDereferencerRequest): ContentD
     });
 
     try {
-        const parsed = parseContent(content, contentType);
+        const parsed = parseContent<PromptTemplate>(content, contentType);
         if (!parsed) {
             return {
                 typedContent: {
