+++ app/src/test/java/io/apicurio/registry/noprofile/rest/v3/PromptRenderTest.java
@@ -26,7 +26,6 @@ import java.util.Map;
 import java.util.UUID;
 
 import static io.restassured.RestAssured.given;
-import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.notNullValue;
@@ -34,6 +33,9 @@ import static org.hamcrest.Matchers.notNullValue;
 /**
  * Integration tests for the prompt template rendering endpoint.
  * POST /groups/{groupId}/artifacts/{artifactId}/versions/{version}/render
+ *
+ * Note: Detailed validation logic tests are in PromptRenderingServiceTest.
+ * These tests focus on API contract, response format, and error handling.
  */
 @QuarkusTest
 public class PromptRenderTest extends AbstractResourceTestBase {
@@ -126,95 +128,27 @@ public class PromptRenderTest extends AbstractResourceTestBase {
     }
 
     @Test
-    public void testRenderMissingVariablePreserved() throws Exception {
-        String group = "missing-var-" + UUID.randomUUID().toString();
-        String artifactId = "partial-prompt";
-
-        String content = createPromptContent("Hello {{name}}, your role is {{role}}.");
-
-        createArtifact(group, artifactId, PROMPT_TEMPLATE, content, ContentTypes.APPLICATION_JSON);
-
-        Map<String, Object> request = Map.of(
-                "variables", Map.of("name", "Bob")
-                // 'role' not provided
-        );
-
-        given().when()
-                .contentType(CT_JSON)
-                .pathParam("groupId", group)
-                .pathParam("artifactId", artifactId)
-                .pathParam("versionExpression", "branch=latest")
-                .body(request)
-                .post("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/{versionExpression}/render")
-                .then()
-                .statusCode(200)
-                .body("rendered", equalTo("Hello Bob, your role is {{role}}."));
-    }
-
-    @Test
-    public void testRenderWithValidation() throws Exception {
-        String group = "validation-" + UUID.randomUUID().toString();
+    public void testRenderValidationErrorResponseFormat() throws Exception {
+        String group = "validation-format-" + UUID.randomUUID().toString();
         String artifactId = "validated-prompt";
 
         String variablesSchema = """
             {
               "name": { "type": "string", "required": true },
-              "count": { "type": "integer", "minimum": 1, "maximum": 100 },
-              "style": { "type": "string", "enum": ["formal", "casual", "technical"] }
+              "count": { "type": "integer", "minimum": 1 }
             }
             """;
 
-        String content = createPromptWithSchema(
-                "Hello {{name}}, please provide {{count}} items in {{style}} style.",
-                variablesSchema
-        );
+        String content = createPromptWithSchema("Hello {{name}}, count: {{count}}", variablesSchema);
 
         createArtifact(group, artifactId, PROMPT_TEMPLATE, content, ContentTypes.APPLICATION_JSON);
 
-        // Valid request
-        Map<String, Object> validRequest = Map.of(
-                "variables", Map.of(
-                        "name", "Test User",
-                        "count", 50,
-                        "style", "formal"
-                )
-        );
-
-        given().when()
-                .contentType(CT_JSON)
-                .pathParam("groupId", group)
-                .pathParam("artifactId", artifactId)
-                .pathParam("versionExpression", "branch=latest")
-                .body(validRequest)
-                .post("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/{versionExpression}/render")
-                .then()
-                .statusCode(200)
-                .body("rendered", containsString("Test User"))
-                .body("rendered", containsString("50"))
-                .body("rendered", containsString("formal"))
-                .body("validationErrors", hasSize(0));
-    }
-
-    @Test
-    public void testRenderValidationRequiredMissing() throws Exception {
-        String group = "required-" + UUID.randomUUID().toString();
-        String artifactId = "required-prompt";
-
-        String variablesSchema = """
-            {
-              "name": { "type": "string", "required": true }
-            }
-            """;
-
-        String content = createPromptWithSchema("Hello {{name}}!", variablesSchema);
-
-        createArtifact(group, artifactId, PROMPT_TEMPLATE, content, ContentTypes.APPLICATION_JSON);
-
-        // Missing required 'name'
+        // Missing required 'name' and 'count' below minimum
         Map<String, Object> request = Map.of(
-                "variables", Map.of()
+                "variables", Map.of("count", 0)
         );
 
+        // Verify the API returns validation errors in the expected format
         given().when()
                 .contentType(CT_JSON)
                 .pathParam("groupId", group)
@@ -224,184 +158,9 @@ public class PromptRenderTest extends AbstractResourceTestBase {
                 .post("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/{versionExpression}/render")
                 .then()
                 .statusCode(200)
-                .body("validationErrors", hasSize(1))
-                .body("validationErrors[0].variableName", equalTo("name"))
-                .body("validationErrors[0].message", containsString("Required"));
-    }
-
-    @Test
-    public void testRenderValidationTypeMismatch() throws Exception {
-        String group = "type-" + UUID.randomUUID().toString();
-        String artifactId = "typed-prompt";
-
-        String variablesSchema = """
-            {
-              "count": { "type": "integer" }
-            }
-            """;
-
-        String content = createPromptWithSchema("Count: {{count}}", variablesSchema);
-
-        createArtifact(group, artifactId, PROMPT_TEMPLATE, content, ContentTypes.APPLICATION_JSON);
-
-        // Wrong type: string instead of integer
-        Map<String, Object> request = Map.of(
-                "variables", Map.of("count", "not-a-number")
-        );
-
-        given().when()
-                .contentType(CT_JSON)
-                .pathParam("groupId", group)
-                .pathParam("artifactId", artifactId)
-                .pathParam("versionExpression", "branch=latest")
-                .body(request)
-                .post("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/{versionExpression}/render")
-                .then()
-                .statusCode(200)
-                .body("validationErrors", hasSize(1))
-                .body("validationErrors[0].variableName", equalTo("count"))
-                .body("validationErrors[0].expectedType", equalTo("integer"))
-                .body("validationErrors[0].actualType", equalTo("string"));
-    }
-
-    @Test
-    public void testRenderValidationEnumInvalid() throws Exception {
-        String group = "enum-" + UUID.randomUUID().toString();
-        String artifactId = "enum-prompt";
-
-        String variablesSchema = """
-            {
-              "style": { "type": "string", "enum": ["a", "b", "c"] }
-            }
-            """;
-
-        String content = createPromptWithSchema("Style: {{style}}", variablesSchema);
-
-        createArtifact(group, artifactId, PROMPT_TEMPLATE, content, ContentTypes.APPLICATION_JSON);
-
-        // Invalid enum value
-        Map<String, Object> request = Map.of(
-                "variables", Map.of("style", "invalid")
-        );
-
-        given().when()
-                .contentType(CT_JSON)
-                .pathParam("groupId", group)
-                .pathParam("artifactId", artifactId)
-                .pathParam("versionExpression", "branch=latest")
-                .body(request)
-                .post("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/{versionExpression}/render")
-                .then()
-                .statusCode(200)
-                .body("validationErrors", hasSize(1))
-                .body("validationErrors[0].variableName", equalTo("style"))
-                .body("validationErrors[0].message", containsString("allowed values"));
-    }
-
-    @Test
-    public void testRenderValidationRangeBelowMinimum() throws Exception {
-        String group = "range-min-" + UUID.randomUUID().toString();
-        String artifactId = "range-prompt";
-
-        String variablesSchema = """
-            {
-              "count": { "type": "integer", "minimum": 10 }
-            }
-            """;
-
-        String content = createPromptWithSchema("Count: {{count}}", variablesSchema);
-
-        createArtifact(group, artifactId, PROMPT_TEMPLATE, content, ContentTypes.APPLICATION_JSON);
-
-        Map<String, Object> request = Map.of(
-                "variables", Map.of("count", 5)
-        );
-
-        given().when()
-                .contentType(CT_JSON)
-                .pathParam("groupId", group)
-                .pathParam("artifactId", artifactId)
-                .pathParam("versionExpression", "branch=latest")
-                .body(request)
-                .post("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/{versionExpression}/render")
-                .then()
-                .statusCode(200)
-                .body("validationErrors", hasSize(1))
-                .body("validationErrors[0].variableName", equalTo("count"))
-                .body("validationErrors[0].message", containsString("less than minimum"));
-    }
-
-    @Test
-    public void testRenderValidationRangeAboveMaximum() throws Exception {
-        String group = "range-max-" + UUID.randomUUID().toString();
-        String artifactId = "range-prompt";
-
-        String variablesSchema = """
-            {
-              "count": { "type": "integer", "maximum": 100 }
-            }
-            """;
-
-        String content = createPromptWithSchema("Count: {{count}}", variablesSchema);
-
-        createArtifact(group, artifactId, PROMPT_TEMPLATE, content, ContentTypes.APPLICATION_JSON);
-
-        Map<String, Object> request = Map.of(
-                "variables", Map.of("count", 200)
-        );
-
-        given().when()
-                .contentType(CT_JSON)
-                .pathParam("groupId", group)
-                .pathParam("artifactId", artifactId)
-                .pathParam("versionExpression", "branch=latest")
-                .body(request)
-                .post("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/{versionExpression}/render")
-                .then()
-                .statusCode(200)
-                .body("validationErrors", hasSize(1))
-                .body("validationErrors[0].variableName", equalTo("count"))
-                .body("validationErrors[0].message", containsString("greater than maximum"));
-    }
-
-    @Test
-    public void testRenderMultipleValidationErrors() throws Exception {
-        String group = "multi-error-" + UUID.randomUUID().toString();
-        String artifactId = "multi-error-prompt";
-
-        String variablesSchema = """
-            {
-              "name": { "type": "string", "required": true },
-              "count": { "type": "integer", "minimum": 0 },
-              "style": { "type": "string", "enum": ["a", "b"] }
-            }
-            """;
-
-        String content = createPromptWithSchema(
-                "{{name}} wants {{count}} {{style}} items.",
-                variablesSchema
-        );
-
-        createArtifact(group, artifactId, PROMPT_TEMPLATE, content, ContentTypes.APPLICATION_JSON);
-
-        Map<String, Object> request = Map.of(
-                "variables", Map.of(
-                        // name is missing (required)
-                        "count", -5,      // below minimum
-                        "style", "invalid" // not in enum
-                )
-        );
-
-        given().when()
-                .contentType(CT_JSON)
-                .pathParam("groupId", group)
-                .pathParam("artifactId", artifactId)
-                .pathParam("versionExpression", "branch=latest")
-                .body(request)
-                .post("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/{versionExpression}/render")
-                .then()
-                .statusCode(200)
-                .body("validationErrors", hasSize(3));
+                .body("validationErrors", hasSize(2))
+                .body("validationErrors[0].variableName", notNullValue())
+                .body("validationErrors[0].message", notNullValue());
     }
 
     @Test
@@ -485,58 +244,5 @@ public class PromptRenderTest extends AbstractResourceTestBase {
                 .post("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/{versionExpression}/render")
                 .then()
                 .statusCode(404);
-    }
-
-    @Test
-    public void testRenderEmptyVariables() throws Exception {
-        String group = "empty-vars-" + UUID.randomUUID().toString();
-        String artifactId = "static-prompt";
-
-        String content = createPromptContent("This is a static prompt with no variables.");
-
-        createArtifact(group, artifactId, PROMPT_TEMPLATE, content, ContentTypes.APPLICATION_JSON);
-
-        Map<String, Object> request = Map.of(
-                "variables", Map.of()
-        );
-
-        given().when()
-                .contentType(CT_JSON)
-                .pathParam("groupId", group)
-                .pathParam("artifactId", artifactId)
-                .pathParam("versionExpression", "branch=latest")
-                .body(request)
-                .post("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/{versionExpression}/render")
-                .then()
-                .statusCode(200)
-                .body("rendered", equalTo("This is a static prompt with no variables."));
-    }
-
-    @Test
-    public void testRenderSpecialCharacters() throws Exception {
-        String group = "special-chars-" + UUID.randomUUID().toString();
-        String artifactId = "special-prompt";
-
-        String content = createPromptContent("Query: SELECT * FROM {{table}} WHERE id = {{id}}");
-
-        createArtifact(group, artifactId, PROMPT_TEMPLATE, content, ContentTypes.APPLICATION_JSON);
-
-        Map<String, Object> request = Map.of(
-                "variables", Map.of(
-                        "table", "users",
-                        "id", 123
-                )
-        );
-
-        given().when()
-                .contentType(CT_JSON)
-                .pathParam("groupId", group)
-                .pathParam("artifactId", artifactId)
-                .pathParam("versionExpression", "branch=latest")
-                .body(request)
-                .post("/registry/v3/groups/{groupId}/artifacts/{artifactId}/versions/{versionExpression}/render")
-                .then()
-                .statusCode(200)
-                .body("rendered", equalTo("Query: SELECT * FROM users WHERE id = 123"));
     }
 }
