+++ pom.xml
@@ -81,7 +81,6 @@
     <module>java-sdk</module>
     <module>java-sdk-v2</module>
     <module>mcp</module>
-    <module>langchain4j-integration</module>
     <module>operator</module>
     <module>schema-resolver</module>
     <module>serdes/generic/serde-common-avro</module>
